+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * Introduce new `ios_merge_strings_files` action. [#329]
 * Introduce new `buildkite_trigger_build` action. [#333]
+* Introduce new `ios_download_strings_files_from_glotpress` action. [#331]
 
 ### Bug Fixes